+++ src/instructions/RabbitizerInstruction/RabbitizerInstruction.c
@@ -25,7 +25,7 @@ void RabbitizerInstruction_init(RabbitizerInstruction *self, uint32_t word, uint
     self->category = RABBITIZER_INSTRCAT_CPU;
 
     self->flags = 0;
-    RAB_INSTR_FLAGS_SET_disasmAsData(self, RAB_TRINARY_VAL_NONE);
+    RAB_INSTR_FLAGS_SET_r5900DisasmAsData(self, RAB_TRINARY_VAL_NONE);
 }
 
 void RabbitizerInstruction_destroy(UNUSED RabbitizerInstruction *self) {