+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.8.0] - 2023-11-12
+
 ### Added
 
 - Add `flag_r5900DisasmAsData` member to the `Instruction` class.